requirements.txt
python-dotenv==1.0.0
langchain==0.0.137
pinecone-client==2.2.1
gradio==3.39.0
openai==0.27.8
tiktoken==0.4.0